Refactor code, group buttons
