+++ src/components/pages/AddTransactionPage.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import TransactionFormInput from '../TransactionFormInput'
 import ToggleSwitch from '../ToggleSwitch'
-import ReturnToOverviewButton from '../ReturnToOverviewButton'
+import ReturnToOverviewButton from '../buttons/ReturnToOverviewButton'
 import styled from 'styled-components/macro'
 import { postNewTransactionEntry } from '../utils/services'
