+++ src/components/pages/TransactionOverviewPage.js
@@ -7,7 +7,7 @@ import {
   editTransactionEntry,
 } from '../utils/services'
 import { monthlyBudget } from '../common/updateMonthlyBudget'
-import AddTransactionButton from '../AddTransactionButton'
+import AddTransactionButton from '../buttons/AddTransactionButton'
 
 export default function TransactionOverviewPage() {
   const [transactions, setTransactions] = useState([])
